Initial Commit
This is actually not the first commit of this project. This project was
named differently and contained some hardcoded keys, as there was only
a single instance of this project running.

This project began as a prototype and was slowly extended to support
more features. This explains the lack of documentation, and the below
average level of quality of certain parts of the project. The goal is
to enhance this project.